Fix problem on tags
If a callable is used, it should be called for each and every instance as the result is supposed to vary between class instances.

An Algolia index has a `tags` class attribute (common to all instances). It should not be redefined in the index model code through an instance (see ``models.AlgoliaIndex.get_raw_record``)
because modifying a class attribute will have an impact on all instances (this is the difference between defining attributes in the class body and in the dunder __init__ method).

See thread here: https://github.com/algolia/algoliasearch-django/issues/228